Need bake test

+++ Assets/Okashi/PermissionManager/Scripts/Editor/PermissionButtonEditor.cs
@@ -71,7 +71,7 @@ namespace Okashi.Permissions.Editors
 
             if (GUILayout.Button("Clear"))
             {
-                t.permissions = new int[0];
+                t.permissions = new ulong[0];
                 OnEnable();
             }
             serializedObject.ApplyModifiedProperties();
